Add exception for query-dimension-records offset
The --offset parameter to query-dimension-records was changed to be silently ignored in DM-45556.  It can't be brought back without going back to the old query system.

We are removing this entirely soon, and it's better to flag the problem to the user than silently ignore what they asked for.

Also added a little more information to a similar information that was being thrown from query-data-ids.