Fix dataset query missing implied dimensions
Revert serialization change from DM-46776 in SerializedDataCoordinate, which caused it to drop values for implied dimensions.

That change broke the transfer of DatasetRefs from server to client, since it drops the values for implied dimensions in the data ID.  Per the documented behavior of dataset queries, the returned refs will have `hasFull() = True` and this behavior is relied on in tutorial notebooks.