DO NOT MERGE: disable caching of collection summaries
With the cache enabled, every query on the server is pulling and processing 17k rows of dataset summaries because filtering is disabled when the cache is on.  (Which is all queries).

This is more like 1 or 60 rows if we use the filter.

The processing of these rows seems to be taking excessive synchronous CPU time (300ms+ on the server).