Do not cache collection summaries by default in queries
We now enable the collection summary cache separately from the collection record cache.  In the Butler.query() context, only the collection record cache is enabled by default.

More often than not, the query context manager is used to execute one query and then exited.  The collection summary cache was often a pessimization in this case -- fetching tens of thousands of rows when only a few were needed.