Introduce special syntax for binds (DM-38497)
Expression parser adds a new syntax which specifies bind identifiers.
Those identifiers are the names prefixed with colon, e.g. `:bind_name`.
The names are presumed to be bind names, as opposed to usual identifiers
that could be either bind names or any other sort of identifiers.

Eventually we want to deprecate simple identifier use for binds, for now
there are no deprecation warnings issued for old syntax.