Reduce queries in Butler.transfer_from
Reduce the number of queries required to import datasets into the registry in `Butler.transfer_from()`.  Datasets are now batched by dimension group instead of by dataset type. This is intended to improve the performance of Prompt Processing, which transfers datasets of many types but mostly with the same dimensions.