Simplify FileDataset serialization
Use a simpler serialization for FileDataset that just represents a single FileDataset object.

The initial version of FileDataset serialization attempted to match the export YAML format.  This took a lot of extra code to generate a complicated nested structure.  Users were going to have to write even more code to un-roll the nested loops every time they use this.