Change pruneDatasets to only prune the requested datasets
Rather than emptyTrash always trying to clear the entire trash
table, add the ability to constrain the removal by a specific
set of refs (that generally are the refs that were just pruned).
Also take the opportunity to chunk deletes into 50,000 files
to reduce the size of transactions that could involve a million
files.