Refactor FileDatastore.transfer_from
Split up the logic in `FileDatastore.transfer_from` so that it is in two parts:
1. Retrieve the file information from the source datastore
2. Execute the transfer

Logic specific to pulling file information out of a `ChainedDatastore` was moved into `ChainedDatastore`, instead of being a special case in `FileDatastore.transfer_from`.

This change will make it easier to provide the file information needed for transferring files from RemoteButler.