Further split up transfer_from logic
Divide the source file lookup logic for transfer_from into two pieces:
1. The portion that only uses the DB
2. The portion that searches the filesystem for additional artifacts

This will let the Butler server only use the DB portion, since it is neither secure or performant to have it search the filesystem for missing files.  It also lets us transfer only Dataset UUIDs from the client to the server, instead of the full refs.

As a side effect, this saves queries to the database for ChainedDatastore.  Previously it was fetching StoredFileInfo objects twice for each child datastore -- now it only does it once.