Reorganize DimensionDataAttacher loop
Re-organize the main loop for DimensionDataAttacher so that it does the complete lookup for each dimension element before moving on to the next one.

This prepares for an upcoming commit that adds support for looking up the implied data ID values if the DataCoordinate is missing them.  That will require the dimension records for the "implying" dimensions to be fetched before the "implied" dimension is looked up.