Allow missing dimension records in import/ingest
Restore the behavior that existed prior to the previous commit, where no error was thrown if missing dimension records could not be found in the database.  Per conversation with the team this was the intended behavior, due to the way visit_detector_region and some other dimensions are populated separately from the dimensions used in the data ID key.