Implement missing registry query shims
Add registry query shims for findDatasets and findRelatedDatasets, which are not yet deprecated but which were not yet implemented in RemoteButler.

Also added a stub for materialize() as a no-op.  The one place we are aware of where this was being used is now gone, but this is not yet deprecated.