Allow ChainDatastore ingest with transfer=None
Fix an issue encountered during integration testing of Prompt Processing with the Butler writer service.  Prompt Processing's target repository uses a ChainedDatastore, and the code was explicitly disallowing ingests with transfer=None.

The exception was probably added to prevent confusion -- normally it would be difficult to split up the files correctly between the multiple roots in a ChainedDatastore.  However, PP is able to get the files into the correct place and the existing code works correctly if the files are where they need to be.