Use unbounded timespan for non-calibration collections.
This was already the expectation for QG calibration lookups and
temporal joins more generally, and it was once the high-level behavior
in QG builds; I'm not sure if we regressed because we removed
special-casing there or because we changed the behavior here.

Happily, fixing this deep in the creation of the dataset_tags
subqueries also addresses the bug fixed on e594fe9b, in a slightly
better (IMO) way.  I've left the extra NULL handling also added on that
commit, though, because it is still possible for other columns to have
NULL timespans.

This also changes what Query.general returns as the timespan for
RUN/TAGGED collections, and I've let that behavior change happen to
keep it consistent with the join behavior.  I don't think its in broad
enough use yet for that to cause trouble.

Unhappily, apparently queryDatasetAssociations has always returned
timespan=None for RUN/TAGGED collections, so that's now special-cased
to continue to do so for backwards compatibility.  It's logically
inconsistent with how temporal joins work, but I'd rather encourage
people to switch to Query.general than deal with fallout from changing
it.