Make ButlerLogRecords extensible.
This extends the ButlerLogRecords serialization format (with backwards
compatibility) to allow subclasses to store additional JSON data with
the logs.  To do so, it makes it so ButlerLogRecords is no longer
itself a Pydantic model, which avoids a few minor existing problems
(it allows inheritance from MutableSequence and lets us make a property
setter for log_format).  It constitutes a minor breaking change for
downstream code, since you now have to call to_json_data() instead of
using Pydantic methods directly to serialize it, but it seems this
class isn't used directly outside middleware.