Fix collection name KeyError in queryDatasetAssociations
Fix a bug where queryDatasetAssociations against a chained collection would fail with `KeyError` when trying to look up the collection name from the `collection_info` dictionary.  This was happening because the associated collections are looked up by the calibration collection name, but we only had an entry in the dictionary for the parent chain.

It also became clear that the `flattenChains` parameter to `queryDatasetAssociation` has no effect, at least since DM-45429 last year.