De-optimize getDatasetRef's DatasetType lookup
Removed an optimization in getDatasetRef that allows it to avoid loading all dataset types from the database.  This optimization will be difficult to maintain in an upcoming commit to support fetching many datasets at the same time.

It would be unusual for an application to call get_dataset() and nothing else -- and even then, it only pays the cost of the fetch once, amid all the other Butler startup costs.  So it doesn't seem worth the cost of maintaining this extra cache update path.