Filter tag collections in DB instead of application
In response to Jim's review comment, join the collections table to filter out the excess rows from tagged collections instead fetching all the rows and filtering in the application.  This will prevent us from loading an excess number of rows for datasets that are tagged in many collections, and should have minimal performance impact because we have to join to get the collection name in either case.